+++ src/ScottPlot4/ScottPlot/PixelPadding.cs
@@ -1,0 +1,18 @@
+﻿namespace ScottPlot
+{
+    public struct PixelPadding
+    {
+        public float Left;
+        public float Right;
+        public float Bottom;
+        public float Top;
+
+        public PixelPadding(float left, float right, float bottom, float top)
+        {
+            Left = left;
+            Right = right;
+            Bottom = bottom;
+            Top = top;
+        }
+    }
+}